Create AttachableDropdown component
